fix(blueprints): address code-review findings (2 security regressions + 4 correctness)
Recall-mode review caught regressions vs master that the migration dropped:

SECURITY:
- patchCanvasItem mass-assignment: restored the PATCHABLE_COLUMNS allowlist (the
  migration's copy predated master's hardening and wrote every param). Now only
  allowlisted columns are written and an empty set returns false.
- ApiCanvas::patch IDOR: restored the project-authorization check master had —
  load item -> its board -> projectId -> isUserAssignedToProject, else 403. Also
  returns 400 (not 500) when no allowlisted fields are present.

CORRECTNESS:
- DelCanvas/DelCanvasItem::post() re-require the 'del' confirmation field (+ id>0);
  the bridge always populates id, so isset($params['id']) alone let any POST delete.
- ShowCanvas::resolveCurrentBoard() casts the board id to int (DB drivers under
  emulated prepares return strings), fixing a strict-types TypeError on the no-id
  render path and the is_int() edit/clone/merge guards silently no-op'ing.
- element.blade.php uses the module-scoped $row['commentCount'] from
  getCanvasItemsById instead of an unscoped per-item countComments() (wrong count
  across modules + N+1).
- showCanvasTop.blade.php guards $statusLabels/$relatesLabels[$filter[...]] so a
  filter persisted under one canvas type doesn't throw "Undefined array key" when
  viewing another type (the filter is a shared session key).

Known/deferred: BlueprintsExport + Goalcanvas Export build XML without escaping
(pre-existing pattern; a correct fix must preserve the rich-text HTML round-trip
the importer relies on — tracked for a dedicated follow-up with import tests).

Verified: PHPStan clean, Pint, Unit 349 tests, Acceptance (blueprints/goals/wiki) green.

+++ app/Domain/Blueprints/Controllers/ShowCanvas.php
@@ -254,7 +254,8 @@ class ShowCanvas extends Controller
         $currentCanvasId = -1;
 
         if (session()->exists($sessionKey)) {
-            $currentCanvasId = session($sessionKey);
+            // Cast: DB drivers (MySQL emulated prepares) return ids as strings.
+            $currentCanvasId = (int) session($sessionKey);
 
             $found = false;
             foreach ($allCanvas as $row) {
@@ -273,7 +274,7 @@ class ShowCanvas extends Controller
         }
 
         if (count($allCanvas) > 0 && session($sessionKey) == '') {
-            $currentCanvasId = $allCanvas[0]['id'];
+            $currentCanvasId = (int) $allCanvas[0]['id'];
             session([$sessionKey => $currentCanvasId]);
         }
